refactor: user - redisUtil 클래스 분리#149
