+++ user-service/src/main/java/com/boeingmerryho/business/userservice/infrastructure/RedisUtilImpl.java
@@ -11,7 +11,10 @@ import org.springframework.stereotype.Component;
 import com.boeingmerryho.business.userservice.application.utils.RedisUtil;
 import com.boeingmerryho.business.userservice.application.utils.jwt.JwtTokenProvider;
 import com.boeingmerryho.business.userservice.domain.User;
+import com.boeingmerryho.business.userservice.exception.ErrorCode;
 
+import io.github.boeingmerryho.commonlibrary.exception.GlobalException;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 
 @Component
@@ -22,6 +25,8 @@ public class RedisUtilImpl implements RedisUtil {
 
 	private static final String USER_INFO_PREFIX = "user:info:";
 	private static final String USER_TOKEN_PREFIX = "user:token:";
+	private static final String BLACKLIST_PREFIX = "blacklist:";
+	private static final String MEMBERSHIP_INFO_PREFIX = "user:membership:info:";
 
 	@Override
 	public void updateUserInfo(User user) {
@@ -57,6 +62,65 @@ public class RedisUtilImpl implements RedisUtil {
 		redisTemplate.expire(tokenKey, expirationTime, TimeUnit.MILLISECONDS);
 
 		return tokenMap;
+	}
+
+	@Override
+	public void clearRedisUserData(Long userId) {
+		redisTemplate.delete(USER_INFO_PREFIX + userId);
+		redisTemplate.delete(USER_TOKEN_PREFIX + userId);
+		redisTemplate.delete(MEMBERSHIP_INFO_PREFIX + userId);
+	}
+
+	@Override
+	public void deleteFromRedisByKey(String tokenKey) {
+		redisTemplate.delete(tokenKey);
+	}
+
+	@Override
+	public void hasKeyInRedis(String key) {
+		if (Boolean.FALSE.equals(redisTemplate.hasKey(key))) {
+			throw new GlobalException(ErrorCode.NOT_FOUND);
+		}
+	}
+
+	@Override
+	public Map<Object, Object> getMapEntriesFromRedis(String key) {
+		return redisTemplate.opsForHash().entries(key);
+	}
+
+	@Override
+	public String getOpsForValue(String key) {
+		return (String)redisTemplate.opsForValue().get(key);
+	}
+
+	@Override
+	public void setOpsForValueRedis(String key, String value) {
+		redisTemplate.opsForValue().set(key, value);
+	}
+
+	@Override
+	public void setTtlAndOpsForValueRedis(String key, String value, Long timeout) {
+		redisTemplate.opsForValue().set(key, value, timeout, TimeUnit.MINUTES);
+	}
+
+	@Override
+	public void blacklistToken(String accessToken) {
+		Claims claims = jwtTokenProvider.parseJwtToken(accessToken);
+		long ttlMillis = jwtTokenProvider.calculateTtlMillis(claims.getExpiration());
+		String blacklistKey = BLACKLIST_PREFIX + accessToken;
+
+		setOpsForValueRedis(blacklistKey, "blacklisted");
+		redisTemplate.expire(blacklistKey, Math.max(ttlMillis, 1), TimeUnit.MILLISECONDS);
+	}
+
+	@Override
+	public void isEqualStoredRefreshToken(Long userId, String refreshToken) {
+		String redisKey = USER_TOKEN_PREFIX + userId;
+		String storedRefreshToken = (String)redisTemplate.opsForHash().get(redisKey, "refreshToken");
+
+		if (storedRefreshToken == null || !storedRefreshToken.equals(refreshToken)) {
+			throw new GlobalException(ErrorCode.JWT_NOT_MATCH);
+		}
 	}
 
 }
